refactor: Messageresolver 제거
